backup/pr2_baseline_20250927_122115/requirements-test.txt
# Testing Requirements for Astron-RPA Integration
# Comprehensive testing dependencies for integration, unit, and UI tests

# Core Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-html>=3.2.0

# HTTP Testing
httpx>=0.24.0
requests>=2.31.0
responses>=0.23.0

# FastAPI Testing
fastapi[all]>=0.100.0
starlette>=0.27.0

# Playwright for UI Testing
playwright>=1.37.0
pytest-playwright>=0.4.0

# Mock and Fixtures
factory-boy>=3.3.0
faker>=19.0.0
freezegun>=1.2.0

# Database Testing
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Async Testing
aiohttp>=3.8.0
aiofiles>=23.0.0
asyncpg>=0.28.0

# Load Testing
locust>=2.16.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.0

# Documentation Testing
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Performance Testing
memory-profiler>=0.61.0
py-spy>=0.3.0

# Security Testing
safety>=2.3.0
semgrep>=1.35.0

# API Testing
tavern>=2.0.0
schemathesis>=3.19.0

# Monitoring and Observability
prometheus-client>=0.17.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.5.0
typer>=0.9.0

# Development Tools
pre-commit>=3.3.0
tox>=4.6.0
nox>=2023.4.22

# Container Testing
testcontainers>=3.7.0
docker>=6.1.0

# Time and Date Testing
python-dateutil>=2.8.0
pytz>=2023.3

# JSON and Data Testing
jsonschema>=4.19.0
cerberus>=1.3.0

# Network Testing
pytest-httpserver>=1.0.0
wiremock>=2.6.0

# Logging Testing
structlog>=23.1.0
loguru>=0.7.0

# Configuration Testing
dynaconf>=3.2.0
python-decouple>=3.8.0

# File System Testing
pytest-tmp-path>=1.0.0
tempfile-utils>=1.0.0

# Parallel Testing
pytest-parallel>=0.1.0
pytest-forked>=1.6.0

# Reporting
allure-pytest>=2.13.0
pytest-json-report>=1.5.0

# Browser Automation (additional)
selenium>=4.11.0
webdriver-manager>=4.0.0

# API Documentation Testing
openapi-spec-validator>=0.6.0
swagger-spec-validator>=3.0.0

# Database Fixtures
pytest-factoryboy>=2.5.0
pytest-fixtures>=0.1.0

# Environment Management
python-multipart>=0.0.6
uvicorn[standard]>=0.23.0

# Debugging
pdb++>=0.10.0
ipdb>=0.13.0
pudb>=2023.1.0

# Performance Profiling
line-profiler>=4.1.0
cProfile-utils>=1.0.0

# Network Mocking
responses>=0.23.0
httpretty>=1.1.0
vcrpy>=5.1.0

# Time Mocking
time-machine>=2.12.0
freezegun>=1.2.0

# Random Data Generation
mimesis>=11.1.0
hypothesis>=6.82.0

# Configuration Validation
marshmallow>=3.20.0
voluptuous>=0.13.0

# CLI Testing
click-testing>=0.1.0
typer-testing>=0.1.0

# File Format Testing
openpyxl>=3.1.0
python-docx>=0.8.0
PyPDF2>=3.0.0

# Image Testing (for screenshot validation)
Pillow>=10.0.0
opencv-python>=4.8.0

# Network Protocol Testing
scapy>=2.5.0
paramiko>=3.3.0

# Encryption Testing
cryptography>=41.0.0
bcrypt>=4.0.0

# JWT Testing
PyJWT>=2.8.0
python-jose>=3.3.0

# OAuth Testing
authlib>=1.2.0
oauthlib>=3.2.0

# WebSocket Testing
websockets>=11.0.0
socketio>=5.8.0

# Message Queue Testing
celery>=5.3.0
kombu>=5.3.0

# Cache Testing
redis>=4.6.0
memcached>=1.6.0

# Search Testing
elasticsearch>=8.9.0
whoosh>=2.7.0

# Monitoring Testing
statsd>=4.0.0
datadog>=0.47.0

# Cloud Testing
boto3>=1.28.0
google-cloud-storage>=2.10.0

# Kubernetes Testing
kubernetes>=27.2.0
kubectl>=1.28.0

# Terraform Testing
python-terraform>=0.10.0

# Infrastructure Testing
ansible>=8.3.0
fabric>=3.2.0

# API Gateway Testing
kong>=0.1.0
nginx>=1.0.0

# Service Mesh Testing
istio>=1.0.0
consul>=1.0.0

# Observability Testing
jaeger-client>=4.8.0
zipkin>=0.2.0

# Log Aggregation Testing
fluentd>=1.0.0
logstash>=1.0.0

# Metrics Testing
influxdb>=5.3.0
grafana-api>=1.0.0

# Alerting Testing
pagerduty>=2.0.0
slack-sdk>=3.21.0

# Backup Testing
duplicity>=2.0.0
restic>=1.0.0

# Disaster Recovery Testing
chaos-engineering>=1.0.0
litmus>=1.0.0
